Move all stdlib atoms into stdlib_minimal
