+++ lib/src/metta/runner/mod.rs
@@ -85,10 +85,9 @@ use std::sync::{Arc, Mutex, OnceLock};
 mod environment;
 pub use environment::{Environment, EnvBuilder};
 
-#[macro_use]
-pub mod stdlib_old;
 use super::interpreter::{interpret, interpret_init, interpret_step, InterpreterState};
 
+#[macro_use]
 pub mod stdlib_minimal;
 use stdlib_minimal::CoreLibLoader;
